Add Linux USBMon dissector
This adds a dissector for USB packets as available from Linux USBMon interface
and a pcap library interface to it. The support is not complete, given the
complexity of the USB specification. Most common values are understood though,
the dissector is fairly useful at this point.

Encoder is not done yet.
A test suite for the decoder is present.